Use numeric basePrice and correct factory doc
Allow decimal prices by changing basePrice validation from integer to numeric in StoreEventsRequest and UpdateEventsRequest. Also fix the HasFactory phpdoc in Event model from EventsFactory to EventFactory. Modified files: backend/app/Http/Requests/StoreEventsRequest.php, backend/app/Http/Requests/UpdateEventsRequest.php, backend/app/Models/Event.php.

+++ backend/app/Http/Requests/UpdateEventsRequest.php
@@ -32,7 +32,7 @@ class UpdateEventsRequest extends FormRequest
             'eventDate' => ['sometimes','date'],
             'eventEndDate' => ['sometimes','date','after_or_equal:eventDate'],
             'category' => ['sometimes','in:cultural,music,sport'],
-            'basePrice' => ['sometimes','integer','min:0'],
+            'basePrice' => ['sometimes','numeric','min:0'],
             'imageUrl' => ['sometimes','url'],
         ];
     }
